Numeric.Probability.Example.Boys

Description

Consider a family of two children. Given that there is a boy in the family, what is the probability that there are two boys in the family?

Documentation

type Probability = Rational

type Dist a = T Probability a

data Child

Constructors

Boy

Girl

type Family = (Child, Child)

birth :: Dist Child

family :: Dist Family

allBoys :: Event Family

existsBoy :: Event Family

familyWithBoy :: Dist Family

twoBoys :: Probability

countBoy :: Child -> Int

countBoys :: Family -> Int

numBoys :: Dist Int