Numeric.Probability.Example.DiceAccum

Description

We play the following game: We roll a die until we stop or we get three spots. In the first case we own all spots obtained so far, in the latter case we own nothing.

What is the strategy for maximizing the expected score?

Documentation

type Score = Int

die :: Fractional prob => T prob Die

roll :: Fractional prob => T prob (Maybe Score)

continue :: Score -> Bool

strategy :: Fractional prob => T prob (Maybe Score)

optimal strategy

game :: Fractional prob => T prob (Maybe Score)

distribution of the scores that are achieved with the optimal strategy

walk :: Int -> IO (Trace (Maybe Score))