Numeric.Probability.Example.NBoys

Description

Ceneralization of Numeric.Probability.Example.Boys

Consider a family of n children. Given that there are k boys in the family, what is the probability that there are m boys in the family?

Documentation

type Family = [Child]

family :: Int -> Dist Family

countBoys :: Family -> Int

boys :: Int -> Event Family

nBoys :: Int -> Int -> Int -> Probability

numBoys :: Int -> Int -> Dist Int

Special cases

onlyBoys1 :: Int -> Probability

only boys in a family that has one boy