Numeric.Probability.Example.Profession

Description

The newspaper Sueddeutsche asked their readers what professions 16 persons have, by only showing the photographies of them and three choices.

Their statistics was: 22% readers had 0 to 5 correct answers (category 0) 75% readers had 6 to 11 correct answers (category 1) 3% readers had 12 to 16 correct answers (category 2)

Can this statistics be explained with random guessing, or is there some information in the photographies that the readers could utilize?

I got 6 correct answers.

Documentation

type Probability = Double

type Dist a = T Probability a

correctAnswers :: Dist Int

categories :: Dist Int