Numeric.Probability.Simulation

Description

Simulation

Documentation

class C c where

Simulation means to repeat a Rnd.change change many times and to accumulate all results into a distribution. Therefore, simulation can be regarded as an approximation of distributions through randomization.

The Sim class allows the overloading of simulation for different kinds of generators, namely transitions and Rnd.change changes:

Trans   a = a -> Dist a   ==>   c = Dist

RChange a = a -> Rnd.T a  ==>   c = Rnd.T = IO

Methods

(~.) :: (Fractional prob, Ord prob, Random prob, Ord a) => Int -> (a -> c a) -> Transition prob a infix 6

returns the final randomized transition

(~..) :: (Fractional prob, Ord prob, Random prob, Ord a) => (Int, Int) -> (a -> c a) -> RExpand prob a infix 6

returns the whole trace for a k-fold simulation

(~*.) :: (Fractional prob, Ord prob, Random prob, Ord a) => (Int, Int) -> (a -> c a) -> Transition prob a infix 8

returns the whole trace for a single simulation