Numeric.Probability.Trace

Description

Tracing

traces of distributions

type Trace a = [a]

type Walk a = a -> Trace a

type Space prob a = Trace (T prob a)

type Expand prob a = a -> Space prob a

walk :: Int -> Change a -> Walk a

walk is a bounded version of the predefined function iterate

traces of random experiments

type RTrace a = T (Trace a)

type RWalk a = a -> RTrace a

type RSpace prob a = T (Space prob a)

type RExpand prob a = a -> RSpace prob a

merge :: (Fractional prob, Ord a) => [RTrace a] -> RSpace prob a

merge converts a list of RTraces into a list of randomized distributions, i.e., an RSpace, by creating a randomized distribution for each list position across all traces

zipListWith :: ([a] -> b) -> [[a]] -> [b]