Numeric.Probability.Transition

Description

Deterministic and probabilistic generators

Transitions

type Change a = a -> a

deterministic generator

type T prob a = a -> T prob a

probabilistic generator

id :: Num prob => T prob a

map :: (Num prob, Ord a) => Change a -> T prob a -> T prob a

map maps a change function to the result of a transformation (map is somehow a lifted form of map) The restricted type of f results from the fact that the argument to t cannot be changed to b in the result T type.

unfold :: (Num prob, Ord a) => T prob (T prob a) -> T prob a

unfold a distribution of transitions into one transition

NOTE: The argument transitions must be independent

compose :: (Num prob, Ord a) => [T prob a] -> T prob a

Composition of transitions similar to compose but with intermediate duplicate elimination.

untilLeft :: (Num prob, Ord a, Ord b) => (a -> T prob (Either b a)) -> T prob a -> T prob b

fix :: (Num prob, Ord a, Ord b) => ((a -> EitherT a (T prob) b) -> a -> EitherT a (T prob) b) -> T prob a -> T prob b

In fix $ go a -> do ...; go xy any action after a go is ignored.

Spreading changes into transitions

type SpreadC prob a = [Change a] -> T prob a

functions to convert a list of changes into a transition

apply :: Num prob => Change a -> T prob a

maybe :: Num prob => prob -> Change a -> T prob a

lift :: Spread prob a -> SpreadC prob a

uniform :: Fractional prob => SpreadC prob a

linear :: Fractional prob => SpreadC prob a

normal :: Floating prob => SpreadC prob a

enum :: RealFloat prob => [Int] -> SpreadC prob a

relative :: RealFloat prob => [prob] -> SpreadC prob a

Spreading transitions into transitions

type SpreadT prob a = [T prob a] -> T prob a

functions to convert a list of transitions into a transition

liftT :: (Num prob, Ord a) => Spread prob (T prob a) -> SpreadT prob a

uniformT :: (Fractional prob, Ord a) => SpreadT prob a

linearT :: (Fractional prob, Ord a) => SpreadT prob a

normalT :: (Floating prob, Ord a) => SpreadT prob a

enumT :: (RealFloat prob, Ord a) => [Int] -> SpreadT prob a

relativeT :: (RealFloat prob, Ord a) => [prob] -> SpreadT prob a