Numeric.Probability.Visualize

Documentation

data FigureEnv

global settings for one figure

Constructors

FE

Fields

fileName :: String

title :: String

xLabel :: String

yLabel :: String

figure :: FigureEnv

default settings for figure environment

types to represent settings for individual plots

data Color

Constructors

Black

Blue

Green

Red

Brown

Gray

Purple

DarkGray

Cyan

LightGreen

Magenta

Orange

Yellow

White

Custom Int Int Int

data LineStyle

Constructors

Solid

Dashed

Dotted

DotDash

LongDash

TwoDash

type PlotFun = Float -> Float

data Plot

settings for individual plots

Constructors

Plot

Fields

ys :: [Float]

xs :: [Float]

color :: Color

lineStyle :: LineStyle

lineWidth :: Int

label :: String

plot :: Plot

default plotting environment

colors :: [Color]

setColor :: Plot -> Color -> Plot

autoColor :: [Plot] -> [Plot]

plotD :: ToFloat a => Dist a -> Plot

create a plot from a distribution

plotRD :: ToFloat a => RDist a -> IO Plot

plotF :: (FromFloat a, ToFloat b) => (Float, Float, Float) -> (a -> b) -> Plot

create a plot from a function

plotL :: ToFloat a => [a] -> Plot

create a plot from a list

plotRL :: ToFloat a => T [a] -> IO Plot

yls :: [Float] -> Plot -> Plot

metaTuple :: [Float] -> [(Float, Float)] -> [(Float, Float)]

incr :: (Ord a, Fractional a) => a -> a

we want to increase the bounds absolutely, account for negative numbers

decr :: (Ord a, Fractional a) => a -> a

we want to increase the bounds absolutely, account for negative numbers

type Vis = IO ()

Visualization output

creating figures

fig :: [Plot] -> Vis

figP :: FigureEnv -> [Plot] -> Vis

showParams :: Show a => [a] -> [String] -> String

legend :: Float -> Float -> [Plot] -> String

drawy :: ToFloat a => Int -> Plot -> [a] -> String

vec :: Show a => [a] -> String

out0 :: FilePath -> String -> IO ()

out1 :: FilePath -> String -> IO ()